interrupt iret

+++ src/arch8/exception.go
@@ -6,12 +6,12 @@ import (
 
 // Excep defines an exception error with a code
 type Excep struct {
-	Code uint32
+	Code byte
 	Err  error
 }
 
 // NewExcep creates a new Exception with a particular code and message.
-func NewExcep(c uint32, s string) *Excep {
+func NewExcep(c byte, s string) *Excep {
 	ret := new(Excep)
 	ret.Code = c
 	ret.Err = errors.New(s)
